+++ src/index.js
@@ -19,7 +19,7 @@ import rules from "./build/rules.js";
 const plugin = {
 	meta: {
 		name: "@eslint/json",
-		version: "0.13.2", // x-release-please-version
+		version: "1.0.1", // x-release-please-version
 	},
 	languages: {
 		json: new JSONLanguage({ mode: "json" }),
@@ -29,6 +29,7 @@ const plugin = {
 	rules,
 	configs: {
 		recommended: {
+			name: "@eslint/json/recommended",
 			plugins: {},
 			rules: recommendedRules,
 		},
@@ -43,3 +44,4 @@ const plugin = {
 export default plugin;
 export { JSONSourceCode };
 export * from "./languages/json-language.js";
+export * from "./types.js";