+++ src/rules/no-duplicate-keys.js
@@ -4,13 +4,18 @@
  */
 
 //-----------------------------------------------------------------------------
+// Imports
+//-----------------------------------------------------------------------------
+
+import { getKey, getRawKey } from "../util.js";
+
+//-----------------------------------------------------------------------------
 // Type Definitions
 //-----------------------------------------------------------------------------
 
 /**
  * @import { MemberNode } from "@humanwhocodes/momoa";
- * @import { JSONRuleDefinition } from "../types.ts";
- *
+ * @import { JSONRuleDefinition } from "../types.js";
  * @typedef {"duplicateKey"} NoDuplicateKeysMessageIds
  * @typedef {JSONRuleDefinition<{ MessageIds: NoDuplicateKeysMessageIds }>} NoDuplicateKeysRuleDefinition
  */
@@ -49,23 +54,22 @@ const rule = {
 			},
 
 			Member(node) {
-				const key =
-					node.name.type === "String"
-						? node.name.value
-						: node.name.name;
+				const key = getKey(node);
+				const rawKey = getRawKey(node, context.sourceCode);
 
 				if (keys.has(key)) {
 					context.report({
 						loc: node.name.loc,
 						messageId: "duplicateKey",
 						data: {
-							key,
+							key: rawKey,
 						},
 					});
 				} else {
 					keys.set(key, node);
 				}
 			},
+
 			"Object:exit"() {
 				keys = objectKeys.pop();
 			},